fix the bench log
